backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3
python-dotenv==1.0.0
openai>=1.12.0
requests==2.31.0
gunicorn==21.2.0
SQLAlchemy>=2.0.0

# Authentication and OAuth
Authlib==1.3.0
PyJWT==2.8.0
Werkzeug==2.3.7
urllib3==1.26.15  # Pin to 1.26.15 to avoid SSL recursion bug with gevent + Python 3.11

# Celery and Redis for async workflow orchestration
celery[redis]==5.3.4
redis==4.6.0

# Async worker support for SSE/Long-Running connections
gevent>=24.2.1  # Updated for Python 3.13 compatibility

# API Documentation (only used when DEBUG_ENABLED=true)
flasgger==0.9.7.1

psycopg2>=2.9.0
